Move functionality from app to url container
