+++ src/components/UrlContainer/UrlContainer.js
@@ -1,22 +1,39 @@
-import React from 'react';
+import React, { Component } from 'react';
 import './UrlContainer.css';
+import { connect } from 'react-redux';
+import { setUrls } from '../../actions';
+import { getUrls } from '../../apiCalls';
 
-const UrlContainer = props => {
-  const urlEls = props.urls.map(url => {
+export class UrlContainer extends Component {
+
+  componentDidMount() {
+    getUrls()
+      .then(data => this.props.setUrls(data.urls))
+      .catch(err => console.error('Error fetching:', err));
+  }
+
+  render() {
     return (
-      <div className="url">
-        <h3>{url.title}</h3>
-        <a href={url.short_url} target="blank">{url.short_url}</a>
-        <p>{url.long_url}</p>
-      </div>
+      !this.props.urls.length ? <p>No urls yet! Find some to shorten!</p> :
+      <section>
+        {this.props.urls.map(url =>
+          <div className="url">
+            <h3>{url.title}</h3>
+            <a href={url.short_url} target="blank">{url.short_url}</a>
+            <p>{url.long_url}</p>
+          </div>
+        )}
+      </section>
     )
-  });
-
-  return (
-    <section>
-      { urlEls.length ? urlEls : <p>No urls yet! Find some to shorten!</p> }
-    </section>
-  )
+  }
 }
 
-export default UrlContainer;
+export const mapStateToProps = state => ({
+  urls: state.urls
+});
+
+export const mapDispatchToProps = dispatch => ({
+  setUrls: urls => dispatch(setUrls(urls))
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(UrlContainer);
